Explain how to use the EVM package in the README
Change the instructions needed to recreate this EVM package (ie `zos add` operations) for more end-user-oriented guides, aimed at linking the EVM package from a user's project and using it directly. The idea is for a user to easily understand how to make use of this package via ZeppelinOS. Please feel free to disregard this PR if the original intention was to show instructions for developers trying to build an EVM package.